requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib==1.7.4
# passlib[bcrypt]
bcrypt==4.0.1
python-multipart
email-validator
httpx
pydantic
pydantic[email]
python-jose[cryptography]

# pip install bcrypt
# pip install sqlalchemy
# pip install pydantic
# pip install 'pydantic[email]'
# pip install passlib
# pip install 'passlib[bcrypt]'